feat: 연료(Fuel) 시스템 구현 #41
- Fuel 도메인 엔티티, 모델, 리포지토리, DataSource 전체 구현
- SharedPreferences 기반 로컬 영속화 (거래 내역 포함)
- 타이머 세션 완료 시 연료 자동 충전 (30분당 1통)
- FuelGauge, StatusCard 등 UI 위젯 int 타입 전환
- main.dart에 FuelLocalDataSource 초기화 추가

+++ lib/core/widgets/space/status_card.dart
@@ -17,7 +17,7 @@ import 'streak_badge.dart';
 /// StatusCard(
 ///   location: '서울',
 ///   locationFlag: '🇰🇷',
-///   fuel: 3.5,
+///   fuel: 3,
 ///   streakDays: 5,
 /// )
 /// ```
@@ -39,7 +39,7 @@ class StatusCard extends StatelessWidget {
   final String? locationFlag;
 
   /// 현재 연료
-  final double fuel;
+  final int fuel;
 
   /// 스트릭 일수
   final int streakDays;
